+++ Tests/Support/Snapshots/edgecases-change-access-control/Sources/Paths/PathsFake.swift
@@ -24,7 +24,7 @@ extension Paths {
         struct GetParameters {
             var enumQueryStringArray: [EnumQueryStringArray]?
             var enumQueryString: EnumQueryString?
-            var enumQueryInteger: Int?
+            var enumQueryInteger: Int32?
 
             enum EnumQueryStringArray: String, Codable, CaseIterable {
                 case greaterThan = ">"
@@ -37,7 +37,7 @@ extension Paths {
                 case xyz = "(xyz)"
             }
 
-            init(enumQueryStringArray: [EnumQueryStringArray]? = nil, enumQueryString: EnumQueryString? = nil, enumQueryInteger: Int? = nil) {
+            init(enumQueryStringArray: [EnumQueryStringArray]? = nil, enumQueryString: EnumQueryString? = nil, enumQueryInteger: Int32? = nil) {
                 self.enumQueryStringArray = enumQueryStringArray
                 self.enumQueryString = enumQueryString
                 self.enumQueryInteger = enumQueryInteger
@@ -61,13 +61,13 @@ extension Paths {
             /// None
             var integer: Int?
             /// None
-            var int32: Int?
+            var int32: Int32?
             /// None
-            var int64: Int?
+            var int64: Int64?
             /// None
             var number: Double
             /// None
-            var float: Double?
+            var float: Float?
             /// None
             var double: Double
             /// None
@@ -87,7 +87,7 @@ extension Paths {
             /// None
             var callback: String?
 
-            init(integer: Int? = nil, int32: Int? = nil, int64: Int? = nil, number: Double, float: Double? = nil, double: Double, string: String? = nil, patternWithoutDelimiter: String, byte: Data, binary: String? = nil, date: NaiveDate? = nil, dateTime: Date? = nil, password: String? = nil, callback: String? = nil) {
+            init(integer: Int? = nil, int32: Int32? = nil, int64: Int64? = nil, number: Double, float: Float? = nil, double: Double, string: String? = nil, patternWithoutDelimiter: String, byte: Data, binary: String? = nil, date: NaiveDate? = nil, dateTime: Date? = nil, password: String? = nil, callback: String? = nil) {
                 self.integer = integer
                 self.int32 = int32
                 self.int64 = int64
