+++ Tests/Support/Snapshots/edgecases-indent-with-two-width-spaces/Sources/Entities/ArrayTest.swift
@@ -6,10 +6,10 @@ import NaiveDate
 
 public struct ArrayTest: Codable {
   public var arrayOfString: [String]?
-  public var arrayArrayOfInteger: [[Int]]?
+  public var arrayArrayOfInteger: [[Int64]]?
   public var arrayArrayOfModel: [[ReadOnlyFirst]]?
 
-  public init(arrayOfString: [String]? = nil, arrayArrayOfInteger: [[Int]]? = nil, arrayArrayOfModel: [[ReadOnlyFirst]]? = nil) {
+  public init(arrayOfString: [String]? = nil, arrayArrayOfInteger: [[Int64]]? = nil, arrayArrayOfModel: [[ReadOnlyFirst]]? = nil) {
     self.arrayOfString = arrayOfString
     self.arrayArrayOfInteger = arrayArrayOfInteger
     self.arrayArrayOfModel = arrayArrayOfModel
@@ -18,7 +18,7 @@ public struct ArrayTest: Codable {
   public init(from decoder: Decoder) throws {
     let values = try decoder.container(keyedBy: StringCodingKey.self)
     self.arrayOfString = try values.decodeIfPresent([String].self, forKey: "array_of_string")
-    self.arrayArrayOfInteger = try values.decodeIfPresent([[Int]].self, forKey: "array_array_of_integer")
+    self.arrayArrayOfInteger = try values.decodeIfPresent([[Int64]].self, forKey: "array_array_of_integer")
     self.arrayArrayOfModel = try values.decodeIfPresent([[ReadOnlyFirst]].self, forKey: "array_array_of_model")
   }
